Make downloadBinaryFrameworkDefinition private in BinaryInstaller
